Rework lifespan
