+++ src/stac_auth_proxy/__init__.py
@@ -8,13 +8,9 @@ with some internal STAC API.
 
 from .app import configure_app, create_app
 from .config import Settings
-from .lifespan import check_conformance, check_server_health, lifespan
 
 __all__ = [
     "create_app",
     "configure_app",
-    "lifespan",
-    "check_conformance",
-    "check_server_health",
     "Settings",
 ]
